// src/src2/api/models/TaxeResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TaxeResponse = {
    idTaxe?: string;
    nomTaxe?: string;
    calculTaxe?: number;
    actif?: boolean;
    typeTaxe?: string;
    porteTaxe?: string;
    montant?: number;
    positionFiscale?: string;
    createdAt?: string;
    updatedAt?: string;
    organizationId?: string;
    agencyId?: string;
};