// src/src2/api/models/TaxeUpdateRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TaxeUpdateRequest = {
    nomTaxe?: string;
    calculTaxe?: number;
    actif?: boolean;
    typeTaxe?: string;
    porteTaxe?: string;
    montant?: number;
    positionFiscale?: string;
    organizationId?: string;
    agencyId?: string;
};